# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "description: \"This is a sample recipes API. Youcan find out more about the API at\"",
    "title": "Recipes API",
    "contact": {
      "name": "Wagner Ricardo Wagner",
      "email": "wagnerricardonet@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/recipes/{id}": {
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid recipe ID"
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "description": "Get new token in exchange for an old one",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Token is new and doesn't need a refresh"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "description": "Login with username and password",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signIn",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}